integration test msomajo
