+++ src/annotator/test/test_files/input.json
@@ -71,5 +71,9 @@
         "lang": "en",
         "_tagopt": "Additional options for the treetagger to use, default is -token -lemma -sgml -quiet to perform POS and Lemma tagging.",
         "tagopt": "-token -lemma -sgml -quiet"
+    },
+    "somajo_dict": {
+        "split_camel_case": true, 
+        "split_sentences": true
     }
 }
